+++ src/main/java/com/example/demo/DemoApplication.java
@@ -11,14 +11,13 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 
-@MapperScan(basePackages = "com/example/demo/mapper")//com/example/demo/mapper -> com.example.demo.mapper ?
+@MapperScan(basePackages = "com/example/demo/mapper")//@MapperScan(basePackages = "com.example.demo.mapper")
 @SpringBootApplication
-//@EnableAutoConfiguration(exclude={DataSourceAutoConfiguration.class})
-
 @ComponentScan(
         basePackages = {"com.together.board.*"
                         ,"com.together.common.conf" // naver.properties set (objectStorage)
                         ,"com.together.common.utils" // 암호화, 임시 폴더 자동 삭제
+                        ,"com.together.member.*"
         }
 )
 @EnableJpaRepositories({"com.together.board.DAO"
@@ -26,10 +25,11 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EntityScan({"com.together.board.bean"
         ,"com.together.member.bean"})
 @EnableScheduling   //일정 시간이 지나면 임시 이미지 저장 폴더 자동 삭제를 위함
+//@EnableAutoConfiguration(exclude={DataSourceAutoConfiguration.class})
 public class DemoApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(DemoApplication.class, args);
         // 123123
-}
+    }
 }
