+++ src/main/resources/naver.properties
@@ -1,0 +1,4 @@
+ncp.accessKey=BD24AC288B8CA5AD78F7	
+ncp.secretKey=2F60D0C8674E9D06B3FB6502BCB441D81FD30D0C
+ncp.regionName=kr-standard
+ncp.endPoint=https://kr.object.ncloudstorage.com
